Fix build errors

+++ tests/unit/io/GeoJSONReaderTest.cpp
@@ -225,12 +225,12 @@ void object::test<18>
 {
     std::string geojson { "{\"type\":\"Feature\",\"geometry\":{\"type\":\"Point\",\"coordinates\":[-117.0,33.0]}, \"properties\": {\"id\": 1, \"name\": \"one\", \"items\": [1,2,3,4]}}" };
     geos::io::GeoJSONFeatureCollection features(geojsonreader.readFeatures(geojson));
-    ensure_equals(1, features.getFeatures().size());
+    ensure_equals(static_cast<size_t>(1), features.getFeatures().size());
     ensure_equals("POINT (-117.000 33.000)", features.getFeatures()[0].getGeometry()->toText());
     ensure_equals(1.0, features.getFeatures()[0].getProperties()["id"].numberValue);
     ensure_equals("one", features.getFeatures()[0].getProperties()["name"].stringValue);
     std::vector<geos::io::GeoJSONValue> values = features.getFeatures()[0].getProperties()["items"].arrayValue;
-    ensure_equals(4, values.size());
+    ensure_equals(static_cast<size_t>(4), values.size());
     ensure_equals(1.0, values[0].numberValue);
     ensure_equals(2.0, values[1].numberValue);
     ensure_equals(3.0, values[2].numberValue);
@@ -249,7 +249,7 @@ void object::test<19>
         "{\"type\":\"Feature\",\"properties\":{\"id\": 3},\"geometry\":{\"type\":\"Point\",\"coordinates\":[-28.125,39.095]}}"
      "]}" };
     geos::io::GeoJSONFeatureCollection features(geojsonreader.readFeatures(geojson));
-    ensure_equals(3, features.getFeatures().size());
+    ensure_equals(static_cast<size_t>(3), features.getFeatures().size());
     ensure_equals("POLYGON ((87.890 64.923, 76.992 55.178, 102.656 46.558, 115.312 60.413, 94.570 58.447, 87.890 64.923))", features.getFeatures()[0].getGeometry()->toText());
     ensure_equals(1.0, features.getFeatures()[0].getProperties()["id"].numberValue);
     ensure_equals("LINESTRING (1.406 48.690, 41.835 34.016, 22.500 13.923)", features.getFeatures()[1].getGeometry()->toText());
